requirements.txt
aiogram==3.4.1
google-cloud-firestore==2.14.0
python-dotenv==1.0.0
aiofiles==23.2.1
openai==1.12.0